Redux persist changes

+++ setup.sh
@@ -16,7 +16,7 @@ touch local.properties && echo "sdk.dir = /Users/$(whoami)/Library/Android/sdk" 
 # Deleting UIWebView related files from node_modules. Which causes IPA rejection.
 # echo "Deleting UIWebView related files from node_modules"
 cd ../
-rm -f node_modules/react-native/React/Views/RCTWebView.h
-rm -f node_modules/react-native/React/Views/RCTWebView.m
-rm -f node_modules/react-native/React/Views/RCTWebViewManager.h
-rm -f node_modules/react-native/React/Views/RCTWebViewManager.m
+test -f node_modules/react-native/React/Views/RCTWebView.h && rm -f node_modules/react-native/React/Views/RCTWebView.h
+test -f node_modules/react-native/React/Views/RCTWebView.m && rm -f node_modules/react-native/React/Views/RCTWebView.m
+test -f node_modules/react-native/React/Views/RCTWebViewManager.h && rm -f node_modules/react-native/React/Views/RCTWebViewManager.h
+test -f node_modules/react-native/React/Views/RCTWebViewManager.m && rm -f node_modules/react-native/React/Views/RCTWebViewManager.m
